DynamoDB: batch_get_item() now only returns up to 16MB of data
